backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-genai
datasets
weaviate-client
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
pyjwt==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
requests==2.31.0
PyJWT